+++ tests/xinput2_tracker_tests.py
@@ -5,16 +5,17 @@ import subprocess
 from volcorner.rect import Rect
 from volcorner import signals
 from volcorner.x11.xinput2tracker import XInput2MouseTracker
-from .util import SignalReceiver, with_xte, with_xvfb
+from .util import SignalReceiver, with_ui, with_xte, with_xvfb
 
 TEST_AREA = Rect.make(64, 64, 1, 1)
 
 
 @with_xvfb
 @with_xte
-def test_mouse_tracking():
+@with_ui
+def test_mouse_tracking(ui):
     # Track the test region.
-    tracker = XInput2MouseTracker()
+    tracker = XInput2MouseTracker(ui)
     tracker.region = TEST_AREA
     tracker.start()
     try:
@@ -35,22 +36,22 @@ def test_mouse_tracking():
 
 @with_xvfb
 @with_xte
-def test_scroll_grab():
-    tracker = XInput2MouseTracker()
+@with_ui
+def test_scroll_grab(ui):
+    tracker = XInput2MouseTracker(ui)
     tracker.start()
     try:
         tracker.grab_scroll()
-        subprocess.check_call(['xte', 'mousemove 0 0'])  # Wait a bit for the grab to finish by moving the mouse
 
         # Test the scroll up event.
         scroll_up = SignalReceiver(signals.SCROLL_UP)
-        subprocess.check_call(['xte', 'mouseclick 4', 'mousemove 0 0'])  # seem to need mousemove to trigger the event?
+        subprocess.check_call(['xte', 'mouseclick 4'])
         scroll_up.wait(0.5)
         assert scroll_up.received
 
         # Test the scroll down event.
         scroll_down = SignalReceiver(signals.SCROLL_DOWN)
-        subprocess.check_call(['xte', 'mouseclick 5', 'mousemove 0 0'])
+        subprocess.check_call(['xte', 'mouseclick 5'])
         scroll_down.wait(0.5)
         assert scroll_down.received
     finally: