+++ volcorner/qt/qtui.py
@@ -1,49 +1,62 @@
 """Qt user interface."""
 
-from collections import namedtuple
+import asyncio
 import json
 import logging
 import os.path
 from pkg_resources import resource_filename
+import struct
 
 from PyQt5 import QtCore
 from PyQt5.QtCore import Qt
 from PyQt5 import QtGui
 from PyQt5 import QtWidgets
 from PyQt5.QtX11Extras import QX11Info
+from quamash import QEventLoop
+import sip
+import xcffib
+import xcffib.shape
+import xcffib.xfixes
+import xcffib.xproto
+from xcffib._ffi import ffi  # Seems to be no public way to parse an event pointer
 
 import volcorner
 from volcorner.corner import Corner
 from volcorner.rect import Rect
-from volcorner.ui import UI
-from volcorner.x11 import xlib
+from volcorner.ui import XCBUI
 
 _log = logging.getLogger("qtgui")
 
 # Animation duration in ms
 DURATION = 200
 
+# X11 desktop ID for "all desktops"
+ALL_DESKTOPS = -1
 
-class QtUI(UI):
+
+class QtUI(XCBUI):
     """Qt user interface."""
     def __init__(self):
         super().__init__()
         self.app = OverlayApplication()
+        self.xcb_connection = self.app.xcb_connection
+        self._eventFilters = {}
+        self.loaded = False
 
     def load(self):
         self.app.load()
+        self.app.on_update_volume(self.volume)
+        self.app.on_update_rect(self.overlay_rect)
 
-    def run_main_loop(self):
-        self.app.exec_()
+    def set_event_loop(self):
+        loop = QEventLoop(self.app)
+        asyncio.set_event_loop(loop)
 
     def show(self):
         self.app.show_overlay.emit()
 
     def hide(self):
         self.app.hide_overlay.emit()
-
-    def stop(self):
-        self.app.stop.emit()
 
     @property
     def corner(self):
@@ -52,7 +65,7 @@ class QtUI(UI):
     # Can't call super().property.__set__: http://bugs.python.org/issue14965
     @corner.setter
     def corner(self, corner):
-        UI.corner.__set__(self, corner)
+        XCBUI.corner.__set__(self, corner)
         self.app.corner = corner
         self.app.update_transform.emit(corner)
 
@@ -63,7 +76,7 @@ class QtUI(UI):
     # Can't call super().property.__set__: http://bugs.python.org/issue14965
     @overlay_rect.setter
     def overlay_rect(self, overlay_rect):
-        UI.overlay_rect.__set__(self, overlay_rect)
+        XCBUI.overlay_rect.__set__(self, overlay_rect)
         self.app.overlay_rect = overlay_rect
         self.app.update_rect.emit(overlay_rect)
 
@@ -74,14 +87,28 @@ class QtUI(UI):
     # Can't call super().property.__set__: http://bugs.python.org/issue14965
     @volume.setter
     def volume(self, volume):
-        UI.volume.__set__(self, volume)
+        XCBUI.volume.__set__(self, volume)
         self.app.update_volume.emit(volume)
+
+    def install_event_filter(self, event_filter):
+        # Wrap filter function in required class
+        filter_obj = NativeEventFilter(self.xcb_connection, event_filter)
+        self._eventFilters[event_filter] = filter_obj
+        self.app.installNativeEventFilter(filter_obj)
+
+    def remove_event_filter(self, event_filter):
+        # Uninstall previously wrapped filter object
+        try:
+            filter_obj = self._eventFilters[event_filter]
+        except KeyError:
+            _log.warning("Tried to uninstall event filter that's not installed: %r", event_filter)
+            return
+        self.app.removeNativeEventFilter(filter_obj)
 
 
 class OverlayApplication(QtWidgets.QApplication):
     show_overlay = QtCore.pyqtSignal()
     hide_overlay = QtCore.pyqtSignal()
-    stop = QtCore.pyqtSignal()
     update_transform = QtCore.pyqtSignal(Corner)
     update_volume = QtCore.pyqtSignal(float)
     update_rect = QtCore.pyqtSignal(Rect)
@@ -100,19 +127,18 @@ class OverlayApplication(QtWidgets.QApplication):
         self.corner = None
         self.window = None
         self._has_set_advanced_window_state = False
+        self.xcb_connection = self.wrap_connection()
 
-        # Use a queued connection since these can be called from another thread
-        self.show_overlay.connect(self.on_show, Qt.QueuedConnection)
-        self.hide_overlay.connect(self.on_hide, Qt.QueuedConnection)
-        self.stop.connect(self.on_stop, Qt.QueuedConnection)
-        self.update_transform.connect(self.on_update_transform, Qt.QueuedConnection)
-        self.update_volume.connect(self.on_update_volume, Qt.QueuedConnection)
-        self.update_rect.connect(self.on_update_rect, Qt.QueuedConnection)
+        self.show_overlay.connect(self.on_show)
+        self.hide_overlay.connect(self.on_hide)
+        self.update_transform.connect(self.on_update_transform)
+        self.update_volume.connect(self.on_update_volume)
+        self.update_rect.connect(self.on_update_rect)
 
         # TODO: can Qt do 1-bit alpha channel?
         # Qt5 lost isCompositingManagerRunning() until 5.7
         if (hasattr(QX11Info, 'isCompositingManagerRunning') and
-                not QX11Info.isCompositingManagerRunning()):
+                not getattr(QX11Info, 'isCompositingManagerRunning')()):
             _log.warning("Compositing window manager NOT detected!  Translucency will be broken.")
 
     def load(self):
@@ -163,9 +189,6 @@ class OverlayApplication(QtWidgets.QApplication):
     def on_hide(self):
         self.queue_animation(self._animate_hide)
 
-    def on_stop(self):
-        self.quit()
-
     def on_update_transform(self, corner):
         if (self.window is not None) and (self.overlay_rect is not None):
             if corner == Corner.TOP_LEFT:
@@ -182,6 +205,8 @@ class OverlayApplication(QtWidgets.QApplication):
                 self.window.translate(0, 0 - self.overlay_rect.height)
 
     def on_update_volume(self, volume):
+        if self.segments is None:
+            return
         assert 0.0 <= volume <= 1.0
         step = 1.0 / len(self.segments)
         for i, segment in enumerate(self.segments):
@@ -189,9 +214,10 @@ class OverlayApplication(QtWidgets.QApplication):
             segment.value = clamp(0.0, relative_value, 1.0)
 
     def on_update_rect(self, rect):
-        if self.window is not None:
-            self.window.move(rect.x1, rect.y1)
-            self.window.setFixedSize(rect.width, rect.height)
+        if self.window is None:
+            return
+        self.window.move(rect.x1, rect.y1)
+        self.window.setFixedSize(rect.width, rect.height)
 
     def _create_window(self, scene):
         window = QtWidgets.QGraphicsView(scene)
@@ -253,12 +279,23 @@ class OverlayApplication(QtWidgets.QApplication):
         # Only need to set this state once.
         if not self._has_set_advanced_window_state:
             self._has_set_advanced_window_state = True
+            window_id = int(self.window.winId())
+            set_window_desktop(self.xcb_connection, window_id, ALL_DESKTOPS)
+            set_empty_window_shape(self.xcb_connection, self.xcb_connection.xfixes, window_id)
+            self.xcb_connection.flush()
 
-            display = QX11Info.display()
-            window_id = self.window.winId()
+    @staticmethod
+    def wrap_connection():
+        """
+        Wrap the current Qt xcb connection in an xcffib Connection object.
 
-            xlib.move_to_desktop(display, window_id, xlib.ALL_DESKTOPS)
-            xlib.set_empty_window_shape(display, window_id)
+        :return: xcffib object
+        """
+        qt_conn = QX11Info.connection()
+        conn_ptr = sip.unwrapinstance(qt_conn)
+        conn = xcffib.wrap(conn_ptr)
+        conn.xfixes = _load_xfixes(conn)
+        return conn
 
 
 class SegmentObject(QtWidgets.QGraphicsObject):
@@ -354,3 +391,89 @@ def path_to(filename):
 
 def clamp(minimum, value, maximum):
     return max(minimum, min(maximum, value))
+
+
+def set_window_desktop(conn, window_id, desktop):
+    """Set the virtual desktop for a window.
+
+    :param xcffib.Connection conn: XCB connection
+    :param window_id: window to change
+    :param desktop: desktop to set, or ALL_DESKTOPS to show on all desktops
+    """
+    net_wm_desktop = _intern_atom(conn, '_NET_WM_DESKTOP')
+    conn.core.ChangeProperty(xcffib.xproto.PropMode.Replace,
+                             window_id,
+                             net_wm_desktop,
+                             xcffib.xproto.Atom.CARDINAL,
+                             32,
+                             1,
+                             struct.pack('i', desktop),
+                             is_checked=True)
+
+
+def set_empty_window_shape(conn, xfixes, window_id):
+    """Set an empty input shape on a window.
+
+    This will prevent the window from receiving any input.
+
+    :param xcffib.Connection conn: XCB connection
+    :param xcffib.xfixes.xfixesExtension xfixes: XFixes extension
+    :param int window_id: window to change
+    """
+    region_id = conn.generate_id()
+    xfixes.CreateRegion(region_id, 0, [])
+    xfixes.SetWindowShapeRegion(window_id, xcffib.shape.SK.Input, 0, 0, region_id,
+                                is_checked=True)
+
+
+def _intern_atom(conn, name):
+    """Get an atom for a string.
+
+    :param xcffib.Connection: XCB connection
+    :param str name: string to look up
+    :returns: atom
+    :rtype: int
+    """
+    return conn.core.InternAtom(False, len(name), name).reply().atom
+
+
+def _load_xfixes(conn):
+    """Return the XFixes extension, checking for at least version 2.
+
+    :param xcffib.Connection conn: XCB connection
+    :raises ValueError: if a compatible XFixes extension is not present
+    :returns: the XFixes extension
+    :rtype: xcffib.xfixes.xfixesExtension
+    """
+    xfixes_major, xfixes_minor = (2, 0)
+    try:
+        xfixes = conn(xcffib.xfixes.key)
+        reply = xfixes.QueryVersion(xfixes_major, xfixes_minor).reply()
+    except:
+        _log.error("Failed to get XFixes 2", exc_info=True)
+        raise ValueError("XFixes 2 is required.")
+    if reply.major_version < 2:
+        _log.error("Need XFixes 2, but only %d.%d is avaliable", reply.major_version,
+                   reply.minor_version)
+        raise ValueError("XFixes 2 is required.")
+    return xfixes
+
+
+class NativeEventFilter(QtCore.QAbstractNativeEventFilter):
+    def __init__(self, conn, event_filter):
+        super().__init__()
+        self.conn = conn
+        self.event_filter = event_filter
+
+    # Detected method signature is wrong.  Should be:
+    # nativeEventFilter(self, Union[QByteArray, bytes, bytearray], sip.voidptr) -> Tuple[bool, int]
+    # noinspection PyMethodOverriding
+    def nativeEventFilter(self, event_type, message):
+        if event_type != 'xcb_generic_event_t':
+            _log.warning('Unexpected native event type %s', event_type)
+            return False, 0
+        generic_event = ffi.cast('xcb_generic_event_t *', message)
+        event = self.conn.hoist_event(generic_event)
+        result = self.event_filter(event)
+        dummy_result = 0  # Used on windows apparently
+        return bool(result), dummy_result