+++ volcorner/x11/xinput2tracker.py
@@ -1,67 +1,55 @@
 """XInput2 mouse tracker."""
 
-__all__ = ['XInput2MouseTracker']
-
 import logging
 
 import xcffib
 from xcffib.xproto import ButtonPressEvent, EventMask, GeGenericEvent, GrabMode, ModMask
 import xcffib.xinput
+from volcorner.logging import TRACE
 from volcorner.rect import Point
-import volcorner.logging
-from volcorner.pollthread import PollThread, run_on_thread
 from volcorner.tracker import MouseTracker
 
-
+__all__ = ['XInput2MouseTracker']
 _log = logging.getLogger("tracking")
 
 
 class XInput2MouseTracker(MouseTracker):
     """XInput mouse tracker."""
-    def __init__(self):
+    def __init__(self, ui):
+        """
+        Initialize a new XInput2MouseTracker.
+
+        :param volcorner.ui.XCBUI ui: UI to install an event handler and get XCB connection from
+        """
         super().__init__()
-        self._conn = None
+        self._ui = ui
+        self._conn = ui.xcb_connection
         self._root = None
-        self._thread = None
+        self._is_listening = False
 
     def start(self):
-        # Connect to X server and load extensions.
-        self._conn = xcffib.connect()
         self._root = self._conn.setup.roots[0].root
         self._conn.xinput = self._load_xinput()
 
         # Select XInput motion events.
         self._select_motion_events()
 
-        # Process events on a background thread.
-        self._thread = PollThread(
-            target=self._tracker_loop,
-            name="XInput2MouseTracker",
-            daemon=True)
-        self._thread.start()
+        # Listen for input events.
+        self._ui.install_event_filter(self.on_event)
+        self._is_listening = True
 
     def stop(self):
-        # Do nothing if already stopped.
-        if self._thread is None:
+        if not self._is_listening:
             return
 
-        # Wait for the thread to finish.
-        self._thread.stop()
-        self._thread = None
+        self._ui.remove_event_filter(self.on_event)
 
-        # Close the X connection.
-        self._root = None
-        self._conn.disconnect()
-        self._conn = None
-
-    @run_on_thread('_thread')
     def grab_scroll(self):
         # Buttons 4 and 5 are the scroll wheel.
         self._grab_button(4)
         self._grab_button(5)
         self._conn.flush()
 
-    @run_on_thread('_thread')
     def ungrab_scroll(self):
         # Buttons 4 and 5 are the scroll wheel.
         self._ungrab_button(4)
@@ -106,27 +94,14 @@ class XInput2MouseTracker(MouseTracker):
         modifiers = ModMask.Any
         self._conn.core.UngrabButton(button, self._root, modifiers)
 
-    def _tracker_loop(self):
-        """Thread main loop."""
-        fd = self._conn.get_file_descriptor()
-        self._thread.poll(lambda _: self._handle_all_events(), fd)
-
-    def _handle_all_events(self):
-        """Handle all available X events."""
-        while True:
-            event = self._conn.poll_for_event()
-            if event is None:
-                break
-            self._handle_event(event)
-
-    def _handle_event(self, event):
+    def on_event(self, event):
         """Handle an X event."""
         if isinstance(event, GeGenericEvent):
             # GeGenericEvent is an XInput2 pointer event.
             # Update the pointer position.
             pointer = self._conn.core.QueryPointer(self._root).reply()
             self.last_point = Point(pointer.root_x, pointer.root_y)
-            _log.log(volcorner.logging.TRACE, "Pointer event %s", self._last_point)
+            _log.log(TRACE, "Pointer event %s", self._last_point)
         elif isinstance(event, ButtonPressEvent):
             # Button press is a scroll up/down event.
             _log.debug("Button press event %s", event.detail)
@@ -135,4 +110,4 @@ class XInput2MouseTracker(MouseTracker):
             elif event.detail == 5:
                 self.on_scroll_down()
         else:
-            _log.debug("Ignoring event %r", event)
+            _log.log(TRACE, "Ignoring event %s", event)